feat: Method to check if query is valid
Since the provided API does only work with certain search terms, the
'BooksAPI' file does now contain a list of said terms and a method that
will check whether the specified query is a valid one. This way we can
avoid calling the webservice for queries that we know that will return no
results.