feat: Book shelf updated with API
The last commit updated the book list on component's state, but changes
didn't persist in the external server.

Now the 'update' method from the API call is being used to perform the
update, and after changes have been stored, the book list will be reloaded
by using the 'getAll' method.