feat: Book search implemented
The book search box has been migrated to a new component. This new
component will keep two properties into its state: the query and the list
of books returned by the server for said query.

For a better performance, the API is not called unless the user has
introduced one of the available search terms specified in the
'SEARCH_TERMS.md' file. The state is updated with an empty array for any
query that doesn't match any of these terms.

Also, a message including the number of results is displayed right above
the search results.