fix: Blank spaces in search box
The ´trim´ function call we did before updating the ´query´ variable in
the ´BookSearch´ component state was preventing the user from entering
blank spaces anywhere in the search box. This made it impossible for the
user to enter multiple-word search terms, such as _'Web Development'_.

The current commit fixes that issue. The ´trim´ function will only be
called before checking if it's a valid term and sending it over to the
server, but it will no longer be applied to the input's value.

Resolves: #2