feat: New 'Loader' component
Although still not in use in the main code, a new 'Loader' component is
now available into the project. This includes a 'loading' text divided in
multiple 'span' tags, their shape trying to ressemble a book.

These books are moving up and down with an animation, but the current
animation is just a quick test and will be changed. Animating the
'asign-self' property doesn't look like a good idea, there's no transition
when the status changes. Instead, the book jumps to the next status in a
single frame.

I'll replace this animation with something based on the 'margin-top' or
'top' properties, which looks far more nicer.