feat: Display loader while fetching data from API
A loading message is now displayed on screen when an Ajax request is sent
to the books API. This way, the user can easily see that something is
actually happening and that the application is not stuck while he receives
an answer and the UI is updated with the new data.

Although the loader design will probably change when the app gets a new
design, this commit closes #4.