fix: Sample test no longer throwing error
Importing 'BooksAPI' into our 'App' file caused the sample test to fail
because 'localStorage' was not defined in the environment where the tests
run.

In order to fix this, I created a new file for storing mock data that will
be used only by project's unit tests. In this case, I've used it to store a
sample token that will replace the one that was supposedly to be read from
the 'localStorage'. Since it's a constant, we can guarantee the token will
be the same everytime we run the tests.

This commit fixes #1.