fix: Default values for categories and rating count
To prevent undefined errors while displaying the details page if any of
the categories array is missing, an empty array is set before calling the
map function if 'categories' does not exist.

In the case of the rating count, a default value of 0 has been set to
ensure a number is displayed in the UI even if we can't get that info.