design: New icon for search button and...
The "+" icon didn't seem the most appropiate to redirect to a search page where
could not even add anything at all, so I've changed it by a magnifying
glass, which is usually what's seen on a search button and it's easy to
understand.

Along with that, I've also added a little surprise around the book cover
images. Don't bother reading anymore of this commit if you don't
want to be spoiled and find it out by yourself.

...

......

..........

Okay, I just couldn't resist the temptation! I felt the need to add a cute cat
somewhere, and that's it! When the user's cursor moves over a book cover,
a cute kitten shows up and says 'meow'. It's a bit random, I know... But
it's so cute, I like how much it adds to the design. I think adding
some extra personal touch somewhere will benefit the final result of this
project, so this little cutie is making it to the submitted app once
everything's ready.