fix: Set shelf changer select value to 'none'
There was an issue when initializing the selected option in the shelf
selector for a book. The problem was that, if the book had no shelf, the
'none' option wouldn't be highlighted.

The reason for this was that the disabled option with the 'Move to...'
placeholder also held a value 'none'. When we set a value for the 'select'
tag, the first match for 'none' was the disabled option, so the real
'none' option was not appearing active when the select was on screen.

The attribute for this placeholder option has been renamed,  hence this
issue is now solved.