refactor + docs + styles: Header moved to 'App' template
As to avoid having the header code duplicated in all of the page
components, the 'App' component does now contain the base layout for all
of the pages.

Some of the app css styles have been updated to adjust the old pages to
the new header.

Along with this changes, the 'App' component is now properly formatted and
updated with documentation.