test: Add sample json files with API output
A new 'data' directory is now included in 'res'. This new folder will
contain mock data files that can be used from within the unit tests, to
fetch local data instead of connecting to the real API.

The files included are the current webservices response for the get book,
get user books and search books requests.