test: Project dependencies updated with 'enzyme'
In order to develop new tests for the application, the project's
`package.json` file has been updated by adding `enzyme` as a new
dependency.

The React official docs recommend this API to manipulate
the DOM from inside the unit tests, and the examples in Jest docs
are also implemented with Enzyme rather than the official React's Test Utils.

References:

 - https://facebook.github.io/react/docs/shallow-renderer.html
 - https://facebook.github.io/jest/docs/en/tutorial-react.html