refactor: Book shelf selector is now a component
The book shelf selector we use to choose the shelf to which we want our
book to move to is now an stand-alone component. This component will be in
charge of updating the select box and sending the selected shelf back to
the parent element that will perform the actual update operation on the
API. This API call is now done in `App.js`, so we no longer have that code
duplicated in book `BookDetails` and `BookListItem`.