refactor + design: Minor adjustments in design
Along with the review of all of the available css code, I made a couple of
adjustments to the design (such as fixing the search input at the top
of the page, so it doesn't dissappear when the user scrolls down the page
to see the results).

Some templates were also updated to work correctly with the new version of
the styles.

Since the stylesheet is pretty long right now, it may be a good idea to
split it into more smaller parts in future updates.