fix: Bottom margin for book details items
Applying a bottom margin on the flex container with the book's image and
basic information made it look bad on smaller screens, since there was not
enough space between the image and the text.

Now we apply that bottom margin to each of the items in the container, so
there's enough space between them if the screen is not big enough for both
to to fit in the same line.