feat: Tags redirect to search page
When the user clicks on any of the tags that are displayed at the bottom
of the book details page, they will be redirected to the search page with
a parameter.

The search input will be initialized with the value of said
route parameter, so they can get a list of the books that matches that
tag.

This parameter is entirely optional: accessing the search page with no
parameters will behave as usual, starting with an empty form.

This commit resolves #5.