fix + refactor: Header logic moved to new component
In order to fix issue #12 I decided to use the 'pathname' property given
by the BrowserRouter component rather than 'window.location.pathname'.

The header buttons are now being hidden when required (back button must
not be visible in the main page, and the info button must dissappear when
the user is in the about page) even in the deployed version of the app in
GitHub pages.

Since there's some logic involved to hide or show those buttons and add a
custom class to the header, I decided to create a new component for the
Header instead of keeping all of this login in 'App.js'.

This closes #12.